+++ Pandora/Paradigm.hs
@@ -1,0 +1,7 @@
+module Pandora.Paradigm (module Exports) where
+
+import Pandora.Paradigm.Structure as Exports
+import Pandora.Paradigm.Junction as Exports
+import Pandora.Paradigm.Inventory as Exports
+import Pandora.Paradigm.Controlflow as Exports
+import Pandora.Paradigm.Basis as Exports