+++ Pandora/Pattern.hs
@@ -1,0 +1,4 @@
+module Pandora.Pattern (module Exports) where
+
+import Pandora.Pattern.Object as Exports
+import Pandora.Pattern.Functor as Exports